tests+verus: prove verification works end-to-end
The new client-side hash and signature verification had no wire-level tests
exercising the reject path. Add MockResolver-backed tests that sync objects
with bad BLAKE3 hashes (dropped), bad delegate signatures (dropped), valid
signatures (accepted), and unsigned objects without delegates (accepted).
Covers the four distinct verification outcomes.

Fix the Verus specs so they actually verify: add main entry point, cast
spec-mode arithmetic to u32, add decreases clause on the hash comparison
loop, use extensional equality for array postcondition, and annotate
quantifier triggers. All 8 specs now verify clean. Wire federation into
the verify-verus nix app so it can be checked with a single command.

Also fix test_full_sync_flow regression from the resource_type change —
the test helper now sets resource_type on all test resources.

+++ crates/aspen-federation/verus/quorum_spec.rs
@@ -15,7 +15,7 @@ pub open spec fn quorum_spec(total: u32) -> u32 {
     if total == 0 {
         1u32
     } else {
-        (total / 2) + 1
+        ((total / 2) + 1) as u32
     }
 }
